Crate mina_merkle::macros::o1_utils

Modules

adjacent_pairs

This module hosts the AdjacentPairs type, which can be used to list all the adjacent pairs of a list. For example, if you have a list of integers [1, 2, 3], you can use it to obtain the list of tuples [(1, 2), (2, 3)].

chunked_polynomial

dense_polynomial

This adds a few utility functions for the [DensePolynomial] arkworks type.

evaluations

This adds a few utility functions for the [Evaluations] arkworks type.

field_helpers

hasher

This module provides the CryptoDigest trait, which provides a generic interface for hashing.

math

This modules implements some math helper functions.

serialization

This adds a few utility functions for serializing and deserializing arkworks types that implement [CanonicalSerialize] and [CanonicalDeserialize].

types

This is used to define common types and associated prologues

Traits

ExtendedDensePolynomial

An extension for the [DensePolynomial] type.

ExtendedEvaluations

An extension for the [Evaluations] type.

FieldHelpers

Field element helpers