Trait mina_merkle::MerkleHasher

pub trait MerkleHasher<const DEGREE: usize = DEFAULT_DEGREE> {
    type Item;
    type Hash;

    fn hash(
        item: &Self::Item, 
        metadata: MerkleTreeNodeMetadata<DEGREE>
    ) -> Self::Hash;
}

Trait for implementing merkle tree hasher degree defaults to 2

Required Associated Types

type Item

Type that MerkleHasher calculates hash from

type Hash

Type that represents the hash value

Required Methods

fn hash(
item: &Self::Item,
metadata: MerkleTreeNodeMetadata<DEGREE>
) -> Self::Hash

Calculates hash from an item and its associated metadata