Trait mina_merkle::MerkleTree

pub trait MerkleTree {
    type Item;
    type Hash;

    fn height(&self) -> u32;
    fn count(&self) -> usize;
    fn root(&mut self) -> Option<Self::Hash>;
    fn add_batch(&mut self, items: impl IntoIterator<Item = Self::Item>);

    fn add(&mut self, item: Self::Item) { ... }
}

Trait for implementing binary merkle tree

Required Associated Types

type Item

Type of the leaf data

type Hash

Type of the hash values

Required Methods

fn height(&self) -> u32

Height of the tree, leaf nodes that store data are counted

fn count(&self) -> usize

Number of leafs

fn root(&mut self) -> Option<Self::Hash>

Root hash, lazy-evaluated

fn add_batch(&mut self, items: impl IntoIterator<Item = Self::Item>)

Adds a batch of leaves in the give order

Provided Methods

fn add(&mut self, item: Self::Item)

Adds a new leaf